chore: add scrap
